requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alabaster==1.0.0
altair==5.4.1
aniso8601==9.0.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asciitree==0.3.3
asttokens==2.4.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
beautifulsoup4==4.12.3
black==19.10b0
blinker==1.8.2
boto3==1.35.49
botocore==1.35.49
braceexpand==0.1.7
cachetools==5.5.0
cdifflib==1.2.6
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.0.2
clip==0.2.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
cytoolz==1.0.0
datasets==2.12.0
decorator==5.1.1
diffusers==0.31.0
dill==0.3.6
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.21.2
editdistance==0.8.1
einops==0.8.0
einops-exts==0.0.4
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.9.0
fasteners==0.19
fasttext==0.9.3
filelock==3.16.1
Flask==2.2.5
Flask-RESTful==0.3.10
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
ftfy==6.3.1
future==1.0.0
g2p-en==2.1.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.67.0
h5py==3.12.1
huggingface-hub==0.15.1
hydra-core==1.3.2
idna==3.10
ijson==3.3.0
imageio==2.36.0
imagesize==1.4.1
importlib_metadata==8.5.0
inflect==7.4.0
iniconfig==2.0.0
intervaltree==3.1.0
ipython==8.29.0
ipywidgets==8.1.5
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jiwer==2.5.2
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_widgets==3.0.13
kaldi-python-io==1.2.2
kaldiio==2.18.0
kiwisolver==1.4.7
kornia==0.7.3
kornia_rs==0.1.5
latexcodec==3.0.0
lazy_loader==0.4
Levenshtein==0.22.0
lhotse==1.27.0
librosa==0.10.2.post1
lightning-utilities==0.11.8
lilcom==1.8.0
llvmlite==0.43.0
loguru==0.7.2
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.1
MarkupSafe==3.0.2
marshmallow==3.23.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
megatron_core==0.5.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.14
narwhals==1.11.0
nemo_text_processing==1.1.0
nemo_toolkit==1.23.0
nerfacc==0.5.3
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numcodecs==0.13.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnx==1.17.0
open_clip_torch==2.28.0
OpenCC==1.1.6
packaging==24.1
pandas==2.2.3
pangu==4.0.6.1
parameterized==0.9.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
plac==1.4.3
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
portalocker==2.10.1
progress==1.6
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyarrow==17.0.0
pybind11==2.13.6
pybtex==0.24.0
pybtex-docutils==1.0.3
pycparser==2.22
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyloudnorm==0.1.1
PyMCubes==0.1.6
pynini==2.1.6.post1
pyparsing==3.2.0
pypinyin==0.53.0
pypinyin-dict==0.8.0
PySocks==1.7.1
pytest==8.3.3
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.0.7
pytz==2024.2
PyYAML==6.0.2
rapidfuzz==2.13.7
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
resampy==0.4.3
responses==0.18.0
rich==13.9.3
rouge_score==0.1.2
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
s3transfer==0.10.3
sacrebleu==2.4.3
sacremoses==0.1.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
sentencepiece==0.2.0
sentry-sdk==2.17.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
sox==1.5.0
soxr==0.5.0.post1
Sphinx==8.1.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-bibtex==2.6.3
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
streamlit==1.39.0
sympy==1.13.1
tabulate==0.9.0
taming-transformers==0.0.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorstore==0.1.45
termcolor==2.5.0
text-unidecode==1.3
textdistance==4.6.3
texterrors==0.5.1
threadpoolctl==3.5.0
timm==1.0.11
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.2
toolz==1.0.0
torch==2.5.0
torchdiffeq==0.2.4
torchmetrics==0.11.0
torchsde==0.2.6
torchvision==0.20.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trampoline==0.1.2
transformers==4.30.0
trimesh==4.5.1
triton==3.1.0
typed-ast==1.5.5
typeguard==4.3.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.18.5
watchdog==5.0.3
wcwidth==0.2.13
webdataset==0.1.62
Werkzeug==3.0.6
wget==3.2
widgetsnbextension==4.0.13
wrapt==1.16.0
xxhash==3.5.0
yarl==1.16.0
youtokentome==1.0.6
zarr==2.18.3
zipp==3.20.2
